+++ pkg/azurecluster/mutate_update.go
@@ -8,12 +8,10 @@ import (
 	"github.com/giantswarm/micrologger"
 	"k8s.io/api/admission/v1beta1"
 	capz "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
-	capzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
 	ctrl "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/azure-admission-controller/internal/patches"
 	"github.com/giantswarm/azure-admission-controller/pkg/generic"
-	"github.com/giantswarm/azure-admission-controller/pkg/key"
 	"github.com/giantswarm/azure-admission-controller/pkg/mutator"
 )
 
@@ -56,31 +54,7 @@ func (m *UpdateMutator) Mutate(ctx context.Context, request *v1beta1.AdmissionRe
 		return []mutator.PatchOperation{}, microerror.Maskf(parsingFailedError, "unable to parse AzureCluster CR: %v", err)
 	}
 
-	patch, err := m.ensureAPIServerLBName(ctx, azureClusterCR)
-	if err != nil {
-		return []mutator.PatchOperation{}, microerror.Mask(err)
-	}
-	if patch != nil {
-		result = append(result, *patch)
-	}
-
-	patch, err = m.ensureAPIServerLBSKU(ctx, azureClusterCR)
-	if err != nil {
-		return []mutator.PatchOperation{}, microerror.Mask(err)
-	}
-	if patch != nil {
-		result = append(result, *patch)
-	}
-
-	patch, err = m.ensureAPIServerLBType(ctx, azureClusterCR)
-	if err != nil {
-		return []mutator.PatchOperation{}, microerror.Mask(err)
-	}
-	if patch != nil {
-		result = append(result, *patch)
-	}
-
-	patch, err = m.ensureAPIServerLBFrontendIPs(ctx, azureClusterCR)
+	patch, err := ensureAPIServerLB(azureClusterCR)
 	if err != nil {
 		return []mutator.PatchOperation{}, microerror.Mask(err)
 	}
@@ -127,40 +101,4 @@ func (m *UpdateMutator) Log(keyVals ...interface{}) {
 
 func (m *UpdateMutator) Resource() string {
 	return "azurecluster"
-}
-
-func (m *UpdateMutator) ensureAPIServerLBName(ctx context.Context, azureCluster *capzv1alpha3.AzureCluster) (*mutator.PatchOperation, error) {
-	if len(azureCluster.Spec.NetworkSpec.APIServerLB.Name) < 1 {
-		return mutator.PatchAdd("/spec/networkSpec/apiServerLB/name", key.APIServerLBName(azureCluster.Name)), nil
-	}
-
-	return nil, nil
-}
-
-func (m *UpdateMutator) ensureAPIServerLBSKU(ctx context.Context, azureCluster *capzv1alpha3.AzureCluster) (*mutator.PatchOperation, error) {
-	if len(azureCluster.Spec.NetworkSpec.APIServerLB.SKU) < 1 {
-		return mutator.PatchAdd("/spec/networkSpec/apiServerLB/sku", key.APIServerLBSKU()), nil
-	}
-
-	return nil, nil
-}
-
-func (m *UpdateMutator) ensureAPIServerLBType(ctx context.Context, azureCluster *capzv1alpha3.AzureCluster) (*mutator.PatchOperation, error) {
-	if len(azureCluster.Spec.NetworkSpec.APIServerLB.Type) < 1 {
-		return mutator.PatchAdd("/spec/networkSpec/apiServerLB/type", key.APIServerLBType()), nil
-	}
-
-	return nil, nil
-}
-
-func (m *UpdateMutator) ensureAPIServerLBFrontendIPs(ctx context.Context, azureCluster *capzv1alpha3.AzureCluster) (*mutator.PatchOperation, error) {
-	if len(azureCluster.Spec.NetworkSpec.APIServerLB.FrontendIPs) < 1 {
-		frontendIPs := []capzv1alpha3.FrontendIP{
-			{Name: key.APIServerLBFrontendIPName(azureCluster.Name)},
-		}
-
-		return mutator.PatchAdd("/spec/networkSpec/apiServerLB/frontendIPs", frontendIPs), nil
-	}
-
-	return nil, nil
 }